[ASV-1589] Moved MoPub initialization to AptoideApplication to start it as early as possible
